+++ R/counts.R
@@ -3,12 +3,12 @@
 #' These functions count the number of correct responses. [countcorrect()]
 #' counts the correct responses regardless of errors, [countcorrect2()]
 #' subtracts the number of errors from number of correct responses,
-#' [sumweighted()] counts the correct responses by giving a weight for different
+#' [sumweighted()] counts the correct responses .by giving a weight for different
 #' responses, [sumscore()] adds up the score for each response.
 #'
 #' @name counts
-#' @templateVar by low
-#' @templateVar vars_input TRUE
+#' @templateVar .by low
+#' @templateVar .input TRUE
 #' @template params-template
 #' @return A [tibble][tibble::tibble-package] contains following values:
 #'   \item{nc}{Count of correct responses. For [countcorrect()].}
@@ -21,34 +21,34 @@ NULL
 
 #' @rdname counts
 #' @export
-countcorrect <- function(data, by, vars_input) {
-  if (is.character(data[[vars_input[["name_acc"]]]])) {
+countcorrect <- function(data, .by, .input) {
+  if (is.character(data[[.input[["name_acc"]]]])) {
     # character input uses "-" to separate individual responses
     data <- data |>
       mutate(
-        r"({vars_input[["name_acc"]]})" := parse_char_resp(
-          .data[[vars_input[["name_acc"]]]]
+        r"({.input[["name_acc"]]})" := parse_char_resp(
+          .data[[.input[["name_acc"]]]]
         )
       ) |>
-      unnest(.data[[vars_input[["name_acc"]]]])
+      unnest(.data[[.input[["name_acc"]]]])
   }
   data |>
-    group_by(across(all_of(by))) |>
+    group_by(across(all_of(.by))) |>
     summarise(
       # NA might be produced in parsing characters
-      nc = sum(.data[[vars_input[["name_acc"]]]] == 1, na.rm = TRUE),
+      nc = sum(.data[[.input[["name_acc"]]]] == 1, na.rm = TRUE),
       .groups = "drop"
     )
 }
 
 #' @rdname counts
 #' @export
-countcorrect2 <- function(data, by, vars_input) {
+countcorrect2 <- function(data, .by, .input) {
   data |>
-    group_by(across(all_of(by))) |>
+    group_by(across(all_of(.by))) |>
     summarise(
       nc_cor = sum(
-        .data[[vars_input[["name_nc"]]]] - .data[[vars_input[["name_ne"]]]]
+        .data[[.input[["name_nc"]]]] - .data[[.input[["name_ne"]]]]
       ),
       .groups = "drop"
     )
@@ -56,13 +56,13 @@ countcorrect2 <- function(data, by, vars_input) {
 
 #' @rdname counts
 #' @export
-sumweighted <- function(data, by, vars_input) {
+sumweighted <- function(data, .by, .input) {
   data |>
-    group_by(across(all_of(by))) |>
+    group_by(across(all_of(.by))) |>
     summarise(
       nc_weighted = sum(
-        .data[[vars_input[["name_weight"]]]] *
-          (.data[[vars_input[["name_acc"]]]] == 1)
+        .data[[.input[["name_weight"]]]] *
+          (.data[[.input[["name_acc"]]]] == 1)
       ),
       .groups = "drop"
     )
@@ -70,11 +70,11 @@ sumweighted <- function(data, by, vars_input) {
 
 #' @rdname counts
 #' @export
-sumscore <- function(data, by, vars_input) {
+sumscore <- function(data, .by, .input) {
   data |>
-    group_by(across(all_of(by))) |>
+    group_by(across(all_of(.by))) |>
     summarise(
-      nc_score = sum(as.numeric(.data[[vars_input[["name_score"]]]])),
+      nc_score = sum(as.numeric(.data[[.input[["name_score"]]]])),
       .groups = "drop"
     )
 }