+++ R/igt.R
@@ -4,20 +4,20 @@
 #' [wikipedia](https://en.wikipedia.org/wiki/Iowa_gambling_task). This modified
 #' version uses pools to simulate cards, but the essential ideas are the same.
 #'
-#' @templateVar by low
-#' @templateVar vars_input TRUE
+#' @templateVar .by low
+#' @templateVar .input TRUE
 #' @template params-template
 #' @return A [tibble][tibble::tibble-package] contains following values:
 #'   \item{sum_outcome}{The total outcome over all trials.}
 #'   \item{perc_good}{The number of choices on "good" pools.}
 #' @export
-igt <- function(data, by, vars_input) {
+igt <- function(data, .by, .input) {
   data |>
-    group_by(across(all_of(by))) |>
+    group_by(across(all_of(.by))) |>
     summarise(
-      sum_outcome = sum(.data[[vars_input[["name_outcome"]]]]),
+      sum_outcome = sum(.data[[.input[["name_outcome"]]]]),
       # good pools have label of "a" and "b"
-      perc_good = mean(.data[[vars_input[["name_pool"]]]] %in% c("a", "b")),
+      perc_good = mean(.data[[.input[["name_pool"]]]] %in% c("a", "b")),
       .groups = "drop"
     )
 }