+++ R/stopsignal.R
@@ -2,8 +2,8 @@
 #'
 #' A classical test on inhibition skills.
 #'
-#' @templateVar by low
-#' @templateVar vars_input TRUE
+#' @templateVar .by low
+#' @templateVar .input TRUE
 #' @template params-template
 #' @return A [tibble][tibble::tibble-package] with the following variables:
 #'   \item{pc_all}{Percent of correct for all the responses.}
@@ -11,39 +11,39 @@
 #'   \item{medrt_go}{Median reaction time (ms) of go trials.}
 #'   \item{ssrt}{Stop signal reaction time (ms).}
 #' @export
-stopsignal <- function(data, by, vars_input) {
+stopsignal <- function(data, .by, .input) {
   data_cor <- data |>
     mutate(
       # remove rt of 100 or less for go trials
       rt_cor = ifelse(
-        .data[[vars_input[["name_rt"]]]] <= 100 &
-          .data[[vars_input[["name_type"]]]] == "go",
-        NA, .data[[vars_input[["name_rt"]]]]
+        .data[[.input[["name_rt"]]]] <= 100 &
+          .data[[.input[["name_type"]]]] == "go",
+        NA, .data[[.input[["name_rt"]]]]
       )
     )
   indices_from_acc <- data_cor |>
-    group_by(across(all_of(by))) |>
+    group_by(across(all_of(.by))) |>
     summarise(
-      pc_all = mean(.data[[vars_input[["name_acc"]]]] == 1),
+      pc_all = mean(.data[[.input[["name_acc"]]]] == 1),
       pc_go = sum(
-        .data[[vars_input[["name_acc"]]]] == 1 &
-          .data[[vars_input[["name_type"]]]] == "go"
-      ) / sum(.data[[vars_input[["name_type"]]]] == "go")
+        .data[[.input[["name_acc"]]]] == 1 &
+          .data[[.input[["name_type"]]]] == "go"
+      ) / sum(.data[[.input[["name_type"]]]] == "go")
     ) |>
     ungroup()
   indices_from_rt <- data_cor |>
-    group_by(across(all_of(by))) |>
+    group_by(across(all_of(.by))) |>
     group_modify(
       ~ .calc_ssrt(
         .x,
-        name_type = vars_input[["name_type"]],
-        name_acc = vars_input[["name_acc"]],
+        name_type = .input[["name_type"]],
+        name_acc = .input[["name_acc"]],
         name_rt = "rt_cor",
-        name_ssd = vars_input[["name_ssd"]]
+        name_ssd = .input[["name_ssd"]]
       )
     ) |>
     ungroup()
-  left_join(indices_from_acc, indices_from_rt, by = by)
+  left_join(indices_from_acc, indices_from_rt, by = .by)
 }
 
 .calc_ssrt <- function(data, name_type, name_acc, name_rt, name_ssd) {