+++ R/switch-congruence.R
@@ -9,8 +9,8 @@
 #' calculates both.
 #'
 #' @name switch-congruence
-#' @templateVar by low
-#' @templateVar vars_input TRUE
+#' @templateVar .by low
+#' @templateVar .input TRUE
 #' @template params-template
 #' @return A [tibble][tibble::tibble-package] with the following variables:
 #'
@@ -43,90 +43,90 @@ NULL
 
 #' @rdname switch-congruence
 #' @export
-complexswitch <- function(data, by, vars_input) {
+complexswitch <- function(data, .by, .input) {
   data_cor <- data |>
     mutate(
       # remove rt of 100 or less
       rt_cor = ifelse(
-        .data[[vars_input[["name_rt"]]]] > 100,
-        .data[[vars_input[["name_rt"]]]], NA
+        .data[[.input[["name_rt"]]]] > 100,
+        .data[[.input[["name_rt"]]]], NA
       ),
       type_block = ifelse(
-        .data[[vars_input[["name_switch"]]]] %in% c("", "pure"),
+        .data[[.input[["name_switch"]]]] %in% c("", "pure"),
         "pure", "mixed"
       ),
       type_switch = ifelse(
         .data[["type_block"]] == "pure",
-        .data[[vars_input[["name_task"]]]],
-        .data[[vars_input[["name_switch"]]]]
+        .data[[.input[["name_task"]]]],
+        .data[[.input[["name_switch"]]]]
       )
     )
   # calculate congruence effect
   cong_eff <- calc_cong_eff(
     data_cor,
-    by = by,
-    name_cong = vars_input[["name_cong"]],
-    name_acc = vars_input[["name_acc"]],
+    .by = .by,
+    name_cong = .input[["name_cong"]],
+    name_acc = .input[["name_acc"]],
     name_rt = "rt_cor"
   )
   # calculate switch cost
   switch_cost <- calc_switch_cost(
     data_cor,
-    by = by,
+    .by = .by,
     name_type_block = "type_block",
     name_type_switch = "type_switch",
-    name_acc = vars_input[["name_acc"]],
+    name_acc = .input[["name_acc"]],
     name_rt = "rt_cor"
   )
-  left_join(cong_eff, switch_cost, by = by)
+  left_join(cong_eff, switch_cost, by = .by)
 }
 
 #' @rdname switch-congruence
 #' @export
-congeff <- function(data, by, vars_input) {
+congeff <- function(data, .by, .input) {
   data_cor <- data |>
     mutate(
       # remove rt of 100 or less
       rt_cor = ifelse(
-        .data[[vars_input[["name_rt"]]]] > 100,
-        .data[[vars_input[["name_rt"]]]], NA
+        .data[[.input[["name_rt"]]]] > 100,
+        .data[[.input[["name_rt"]]]], NA
       )
     )
   calc_cong_eff(
     data_cor,
-    by = by,
-    name_cong = vars_input[["name_cong"]],
-    name_acc = vars_input[["name_acc"]],
+    .by = .by,
+    name_cong = .input[["name_cong"]],
+    name_acc = .input[["name_acc"]],
     name_rt = "rt_cor"
   )
 }
 
 #' @rdname switch-congruence
 #' @export
-switchcost <- function(data, by, vars_input) {
+switchcost <- function(data, .by, .input) {
   data_cor <- data |>
     mutate(
       # remove rt of 100 or less
       rt_cor = ifelse(
-        .data[[vars_input[["name_rt"]]]] > 100,
-        .data[[vars_input[["name_rt"]]]], NA
+        .data[[.input[["name_rt"]]]] > 100,
+        .data[[.input[["name_rt"]]]], NA
       ),
       type_block = ifelse(
-        .data[[vars_input[["name_switch"]]]] %in% c("", "pure"),
+        .data[[.input[["name_switch"]]]] %in% c("", "pure"),
         "pure", "mixed"
       ),
       type_switch = ifelse(
         .data[["type_block"]] == "pure",
-        .data[[vars_input[["name_task"]]]],
-        .data[[vars_input[["name_switch"]]]]
+        .data[[.input[["name_task"]]]],
+        .data[[.input[["name_switch"]]]]
       )
     )
   calc_switch_cost(
     data_cor,
-    by = by,
+    .by = .by,
     name_type_block = "type_block",
     name_type_switch = "type_switch",
-    name_acc = vars_input[["name_acc"]],
+    name_acc = .input[["name_acc"]],
     name_rt = "rt_cor"
   )
 }