+++ R/symncmp.R
@@ -4,8 +4,8 @@
 #' time (mrt), distance effect (dist_effect) and adjusted distance effect
 #' (dist_effect_cor).
 #'
-#' @templateVar by low
-#' @templateVar vars_input TRUE
+#' @templateVar .by low
+#' @templateVar .input TRUE
 #' @template params-template
 #' @return A [tibble][tibble::tibble-package] contains following values:
 #'   \item{pc}{Percentage of correct responses.}
@@ -13,18 +13,18 @@
 #'   \item{dist_eff}{Distance effect.}
 #' @seealso [nsymncmp()] for non-symbolic number comparison.
 #' @export
-symncmp <- function(data, by, vars_input) {
+symncmp <- function(data, .by, .input) {
   data_cor <- data |>
     mutate(
       rt_cor = ifelse(
-        .data[[vars_input[["name_rt"]]]] > 100,
-        .data[[vars_input[["name_rt"]]]], NA
+        .data[[.input[["name_rt"]]]] > 100,
+        .data[[.input[["name_rt"]]]], NA
       )
     )
   basics <- calc_spd_acc(
     data_cor,
-    by,
-    name_acc = vars_input[["name_acc"]],
+    .by,
+    name_acc = .input[["name_acc"]],
     name_rt = "rt_cor",
     rt_rtn = "mean",
     acc_rtn = "percent"
@@ -32,7 +32,7 @@ symncmp <- function(data, by, vars_input) {
   fit_errproof <- purrr::possibly(
     ~ stats::coef(stats::lm(
       as.formula(
-        stringr::str_glue(r"({vars_input[["name_rt"]]} ~ dist)")
+        stringr::str_glue(r"({.input[["name_rt"]]} ~ dist)")
       ),
       .x
     ))[["dist"]],
@@ -40,20 +40,20 @@ symncmp <- function(data, by, vars_input) {
   )
   dist_eff <- data_cor |>
     mutate(
-      dist = .data[[vars_input[["name_big"]]]] -
-        .data[[vars_input[["name_small"]]]]
+      dist = .data[[.input[["name_big"]]]] -
+        .data[[.input[["name_small"]]]]
     ) |>
-    group_nest(across(all_of(by))) |>
+    group_nest(across(all_of(.by))) |>
     mutate(
       dist_eff = purrr::map_dbl(
         .data[["data"]],
         ~ .x |>
           filter(
-            .data[[vars_input[["name_acc"]]]] == 1
+            .data[[.input[["name_acc"]]]] == 1
           ) |>
           fit_errproof()
       ),
       .keep = "unused"
     )
-  left_join(basics, dist_eff, by = by)
+  left_join(basics, dist_eff, by = .by)
 }