+++ tests/testthat/test-multisense.R
@@ -10,12 +10,12 @@ data <- expand_grid(
   mutate(rt = rexp(n(), 0.001))
 
 test_that("Default behavior works", {
-  expect_snapshot(preproc(data, multisense, by = "id"))
+  expect_snapshot(preproc(data, multisense, .by = "id"))
 })
 
 test_that("Works with multiple grouping variables", {
   data <- mutate(data, id1 = id + 1)
-  expect_snapshot(preproc(data, multisense, by = c("id", "id1")))
+  expect_snapshot(preproc(data, multisense, .by = c("id", "id1")))
 })
 
 
@@ -25,10 +25,10 @@ test_that("Works when character case is messy", {
       type = recode(type, Image = "image")
     )
   expect_silent(
-    case_messy <- preproc(data_case_messy, multisense, by = "id")
+    case_messy <- preproc(data_case_messy, multisense, .by = "id")
   )
   expect_identical(
     case_messy,
-    preproc(data, multisense, by = "id")
+    preproc(data, multisense, .by = "id")
   )
 })