+++ tests/testthat/test-span.R
@@ -45,12 +45,12 @@ data <- tibble::tibble(
   ungroup()
 
 test_that("Default behavior works", {
-  expect_snapshot(preproc(data, span, by = "id"))
+  expect_snapshot(preproc(data, span, .by = "id"))
 })
 
 test_that("Works with multiple grouping variables", {
   data <- mutate(data, id1 = id + 1)
-  expect_snapshot(preproc(data, span, by = c("id", "id1")))
+  expect_snapshot(preproc(data, span, .by = c("id", "id1")))
 })
 
 test_that("Works when no acc column found", {
@@ -80,7 +80,7 @@ test_that("Works when no acc column found", {
         )
     ) |>
     ungroup()
-  result_no_acc <- preproc(data_no_acc, span, by = "id")
+  result_no_acc <- preproc(data_no_acc, span, .by = "id")
   expect_snapshot(result_no_acc)
   expect_true(all(is.na(result_no_acc$nc)))
   data_repairable <- data_no_acc |>
@@ -92,7 +92,7 @@ test_that("Works when no acc column found", {
         slen, ~ paste(sample(seq_len(.x)), collapse = "-")
       )
     )
-  result_repaired <- preproc(data_repairable, span, by = "id")
+  result_repaired <- preproc(data_repairable, span, .by = "id")
   expect_snapshot(result_repaired)
   expect_true(all(!is.na(result_repaired$nc)))
 })