+++ tests/testthat/test-stopsignal.R
@@ -49,12 +49,12 @@ data <- expand_grid(
   arrange(id, trial)
 
 test_that("Default behavior works", {
-  expect_snapshot(preproc(data, stopsignal, by = "id"))
+  expect_snapshot(preproc(data, stopsignal, .by = "id"))
 })
 
 test_that("Works with multiple grouping variables", {
   data <- mutate(data, id1 = id + 1)
-  expect_snapshot(preproc(data, stopsignal, by = c("id", "id1")))
+  expect_snapshot(preproc(data, stopsignal, .by = c("id", "id1")))
 })
 
 test_that("Works when character case is messy", {
@@ -63,10 +63,10 @@ test_that("Works when character case is messy", {
       type = recode(type, Stop1 = "stop1")
     )
   expect_silent(
-    case_messy <- preproc(data_case_messy, stopsignal, by = "id")
+    case_messy <- preproc(data_case_messy, stopsignal, .by = "id")
   )
   expect_identical(
     case_messy,
-    preproc(data, stopsignal, by = "id")
+    preproc(data, stopsignal, .by = "id")
   )
 })